requirements.txt
mkdocs>=1.5.3
mkdocs-material>=9.5.0
pymdown-extensions>=10.7
mkdocs-minify-plugin>=0.8.0

# Sincronización con Notion
requests>=2.31.0
python-dotenv>=1.0.0

